use script over shell module to properly run scripts on defined hosts
To fully support the playbook to run on non localhosts, the tasks
calling the shell scripts under files/ will need to use the script
module. Without this, the prior delegate_to will continue to run
always on localhost.

Also trims any new lines on the pod name when running tests using
oc exec. Without this when using script module, the second parameter
to the script was never being found.